Refactored local_xxx commands
