+++ client-programs/pkg/cmd/local_config_view_cmd.go
@@ -3,13 +3,13 @@ package cmd
 import (
 	"fmt"
 
+	"github.com/educates/educates-training-platform/client-programs/pkg/config"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
-	"github.com/educates/educates-training-platform/client-programs/pkg/config"
 	"gopkg.in/yaml.v2"
 )
 
-var (
+const (
 	localConfigViewExample = `
   # View local educates cluster configuration by default. Uses nip.io wildcard domain and Kind as provider config defaults
   educates local config view --config NULL
