+++ client-programs/pkg/cmd/local_registry_prune_cmd.go
@@ -7,6 +7,11 @@ import (
 	"github.com/educates/educates-training-platform/client-programs/pkg/registry"
 )
 
+const localRegistryPruneExample = `
+  # Prune the local image registry
+  educates local registry prune
+`
+
 type LocalRegistryPruneOptions struct {
 }
 
@@ -28,6 +33,7 @@ func (p *ProjectInfo) NewLocalRegistryPruneCmd() *cobra.Command {
 		Use:   "prune",
 		Short: "Prunes the local image registry (deletes any untagged image)",
 		RunE:  func(_ *cobra.Command, _ []string) error { return o.Run() },
+		Example: localRegistryPruneExample,
 	}
 
 	return c
