+++ client-programs/pkg/cmd/local_secrets_add_cmd_group.go
@@ -1,0 +1,35 @@
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+	"k8s.io/kubectl/pkg/util/templates"
+)
+
+func (p *ProjectInfo) NewLocalSecretsAddCmdGroup() *cobra.Command {
+	var c = &cobra.Command{
+		Args:  cobra.NoArgs,
+		Use:   "add",
+		Short: "Add secret to the cache",
+	}
+
+	// Use a command group as it allows us to dictate the order in which they
+	// are displayed in the help message, as otherwise they are displayed in
+	// sort order.
+	commandGroups := templates.CommandGroups{
+		{
+			Message: "Available Commands:",
+			Commands: []*cobra.Command{
+				p.NewLocalSecretsAddCaCmd(),
+				p.NewLocalSecretsAddDockerRegistryCmd(),
+				// NewLocalSecretsAddGenericCmd(),
+				p.NewLocalSecretsAddTlsCmd(),
+			},
+		},
+	}
+
+	commandGroups.Add(c)
+
+	templates.ActsAsRootCommand(c, []string{"--help"}, commandGroups...)
+
+	return c
+}
